Refactor based on code-review comments.
