+++ pkg/broker/broker.go
@@ -114,20 +114,21 @@ func NewBroker(
 		)
 	}
 
-	err = b.asyncEngine.RegisterJob("waitForParentStep", b.doWaitForParentStep)
+	err = b.asyncEngine.RegisterJob("checkParentStatus", b.doCheckParentStatus)
 	if err != nil {
 		return nil, errors.New(
-			"error registering async job for executing wait for parent step",
+			"error registering async job for executing check of parent status",
 		)
 	}
 
 	err = b.asyncEngine.RegisterJob(
-		"waitForChildrenStep",
-		b.doWaitForChildrenStep,
+		"checkChildrenStatuses",
+		b.doCheckChildrenStatuses,
 	)
 	if err != nil {
 		return nil, errors.New(
-			"error registering async job for executing wait for parent step",
+			"error registering async job for executing check of children " +
+				"statuses",
 		)
 	}
